+++ pkg/pve/pve_cluster.go
@@ -7,13 +7,25 @@ import (
 	"strconv"
 )
 
+type PVEClusterService struct {
+	api      *PVE
+	Firewall *PVEClusterFirewallService
+}
+
+func newPVEClusterService(api *PVE) *PVEClusterService {
+	service := new(PVEClusterService)
+	service.api = api
+	service.Firewall = newPVEClusterFirewallService(api)
+	return service
+}
+
 // GetNextVMID returns the next available VMID.
-func (api *PVE) GetNextVMID() (int, error) {
+func (s *PVEClusterService) GetNextVMID() (int, error) {
 	method := http.MethodGet
 	path := path.Join("/cluster", "/nextid")
 
 	res := new(string)
-	err := api.httpClient.sendReq(method, path, nil, res)
+	err := s.api.client.sendReq(method, path, nil, res)
 
 	if err != nil {
 		return 0, err