+++ pkg/pve/pve_node_firewall.go
@@ -6,6 +6,16 @@ import (
 	"strconv"
 )
 
+type PVENodeFirewallService struct {
+	api *PVE
+}
+
+func newPVENodeFirewallService(api *PVE) *PVENodeFirewallService {
+	service := new(PVENodeFirewallService)
+	service.api = api
+	return service
+}
+
 type GetNodeFirewallRulesResponse[Position interface{ int | string }] struct {
 	Action          string           `json:"action"`
 	Comment         string           `json:"comment"`
@@ -24,39 +34,39 @@ type GetNodeFirewallRulesResponse[Position interface{ int | string }] struct {
 	Type            string           `json:"type"`
 }
 
-// GetNodeRules retrieves node's firewall rules.
-func (api *PVE) GetNodeRules(node string) ([]GetNodeFirewallRulesResponse[int], error) {
+// GetRules retrieves node's firewall rules.
+func (s *PVENodeFirewallService) GetRules(node string) ([]GetNodeFirewallRulesResponse[int], error) {
 	method := http.MethodGet
 	path := path.Join("/nodes", node, "/firewall/rules")
 
 	res := &[]GetNodeFirewallRulesResponse[int]{}
-	err := api.httpClient.sendReq(method, path, nil, res)
+	err := s.api.client.sendReq(method, path, nil, res)
 
 	return *res, err
 }
 
-// GetNodeRulesByPos Retrieves a single node's firewall rule using rule's position (pos) as an index.
-func (api *PVE) GetNodeRulesByPos(node string, pos int) (GetNodeFirewallRulesResponse[string], error) {
+// GetRulesByPos Retrieves a single node's firewall rule using rule's position (pos) as an index.
+func (s *PVENodeFirewallService) GetRulesByPos(node string, pos int) (GetNodeFirewallRulesResponse[string], error) {
 	method := http.MethodGet
 	path := path.Join("/nodes", node, "/firewall/rules", strconv.Itoa(pos))
 
 	res := &GetNodeFirewallRulesResponse[string]{}
-	err := api.httpClient.sendReq(method, path, nil, res)
+	err := s.api.client.sendReq(method, path, nil, res)
 
 	return *res, err
 }
 
-// ReadNodeLog Retrieves node's firewall log entries.
+// ReadLog Retrieves node's firewall log entries.
 //
 // TODO: Add missing limit, since, start, until parameters shown in [docs].
 //
 // [docs]: https://pve.proxmox.com/pve-docs/api-viewer/#/nodes/{node}/firewall/log
-func (api *PVE) ReadNodeLog(node string) ([]FirewallLogEntry, error) {
+func (s *PVENodeFirewallService) ReadLog(node string) ([]FirewallLogEntry, error) {
 	method := http.MethodGet
 	path := path.Join("/nodes", node, "/firewall/log")
 
 	res := &[]FirewallLogEntry{}
-	err := api.httpClient.sendReq(method, path, nil, res)
+	err := s.api.client.sendReq(method, path, nil, res)
 
 	return *res, err
 }